Moved Command class to init of the package to avoid redundant imports
