+++ pypixelart/command/commands.py
@@ -4,7 +4,7 @@ from typing import Tuple
 
 import pygame as pg
 
-from pypixelart.command.command import Command
+from pypixelart.command import Command
 from pypixelart.constants import ALPHA
 from pypixelart.symmetry_type import SymmetryType
 from pypixelart.utils import draw_pixel
